refactor(services): move single-owner services into their owning domains
Follow-up to #1451 / #1482. Cleans up the top-level `crates/server/src/services/`
layer: any module with a clear single owning domain moves under
`domains/<owner>/`; modules that genuinely have no single owner stay in
`services/` with justification.

Moved:
- `eval_runner.rs` → `domains/evals/runner.rs`
- `virtual_mount_registry.rs` → `domains/session_files/virtual_mount_registry.rs`
- `scoped_mcp.rs` → `domains/mcp_servers/scoped_mcp.rs`
- `leased_resource.rs` → `domains/session_resources/leased_resource.rs`
- `capability_validation.rs` → `domains/capabilities/validation.rs`
- `budget_tests.rs` → `domains/budgets/tests.rs` (also unwraps the redundant
  inner `mod tests { ... }` wrapper)

Kept in `services/` (justified in `specs/domains.md`):
- `capability` — registry in `Ctx`, used by every domain.
- `event` — persistence + fanout; called by multiple emitting domains.
- `llm_resolver` — spans `llm_providers` + `llm_models` + request params; no
  single owner.
- `model_sync` — background listener reconciling provider catalogs.
- `principal` — shared lookup spanning users + agent identities.
- `usage_tracking` — cross-domain event listener feeding budgets.

Also updates `specs/domains.md` with the "shared services" rule (3-part
test) and lists the current kept modules. No behavioral change — purely
move + import rewrite.

Rust tests: 1012 lib + 8 security integration tests pass; pre-push green.
