+++ crates/server/src/domains/capabilities/mod.rs
@@ -6,5 +6,6 @@
 pub mod commands;
 pub mod queries;
 pub mod types;
+pub mod validation;
 
 pub use commands::*;
