+++ crates/server/src/domains/capabilities/validation.rs
