+++ crates/server/src/domains/mcp_servers/scoped_mcp.rs
