+++ crates/server/src/domains/session_files/mod.rs
@@ -1,9 +1,11 @@
-// Session files domain — commands, queries, and types.
+// Session files domain — commands, queries, service, virtual-mount registry.
 
 pub mod commands;
 pub mod queries;
 pub mod service;
 pub mod types;
+pub mod virtual_mount_registry;
 
 pub use commands::*;
 pub use service::*;
+pub use virtual_mount_registry::VirtualMountRegistry;
