+++ crates/server/src/domains/session_files/virtual_mount_registry.rs
