+++ crates/server/src/domains/session_resources/mod.rs
@@ -3,9 +3,11 @@
 // See specs/domains.md for the pattern.
 
 pub mod commands;
+pub mod leased_resource;
 pub mod queries;
 pub mod service;
 pub mod types;
 
 pub use commands::*;
+pub use leased_resource::LeasedResourceService;
 pub use service::*;
