+++ crates/server/src/services/mod.rs
@@ -1,26 +1,31 @@
 // Cross-cutting infrastructure modules.
 //
-// Domain-owned business logic lives under `crate::domains::*`. The remaining
-// top-level `services` modules are infra adapters, validators, listeners, or
-// other shared helpers that do not define a separate user-facing service layer.
+// Domain-owned business logic lives under `crate::domains::*`. The modules
+// that remain here are infra adapters, validators, listeners, or shared
+// helpers with no single owning domain:
+//
+// - `capability` — capability registry threaded through `Ctx`; used by every
+//   domain.
+// - `event` — event persistence + fanout listener; called by domains that
+//   emit events.
+// - `llm_resolver` — spans `llm_providers` + `llm_models` + params; no single
+//   owner.
+// - `model_sync` — background provider-model sync listener.
+// - `principal` — resolves users + agent identities; shared.
+// - `usage_tracking` — cross-domain event listener feeding budgets.
+//
+// Anything with a clear single owner belongs under `domains/<owner>/`. See
+// `specs/domains.md` for the "shared services" rule.
 
-#[cfg(test)]
-mod budget_tests;
 pub mod capability;
-pub(crate) mod capability_validation;
-pub mod eval_runner;
 pub mod event;
-pub mod leased_resource;
 pub mod llm_resolver;
 pub mod model_sync;
 pub mod principal;
-pub mod scoped_mcp;
 pub mod usage_tracking;
-pub mod virtual_mount_registry;
 
 pub use capability::CapabilityService;
 pub use event::EventService;
-pub use leased_resource::LeasedResourceService;
 pub use llm_resolver::{LlmResolverService, ResolvedModel};
 pub use model_sync::{ModelSyncService, SyncResult};
 pub use principal::{PrincipalService, row_to_principal};
