+++ specs/domains.md
@@ -328,15 +328,40 @@ contract. When introducing a new such entity, add:
    (PG + in-memory).
 2. One `inventory::submit!` block in `domains/org_resolver.rs`.
 
-### When NOT to use this pattern
+### Shared services (`crates/server/src/services/`)
 
-- **Internal-only logic** with no user-facing API surface (e.g. `llm_resolver`,
-  `model_sync`, `usage_tracking`). Keep these as standalone services/modules.
-- **Cross-cutting listeners** (e.g. `NotificationEventListener`,
-  `UsageTrackingListener`). These are event-driven, not command-driven.
-- **Complex orchestration** that spans multiple domains (e.g. session creation
-  resolving agent + harness + capabilities). These remain in their domain's
-  commands but may call other domains' queries.
+The `services/` layer exists only for code that has no single domain owner.
+A file belongs there when **all three** are true:
+
+1. No single domain owns it. If one domain is the natural owner, it moves
+   there — even if other domains consume it via free helpers.
+2. No user-facing command depends on it as a direct callee. Commands call
+   domain code; domain code may call shared services. If an HTTP/MCP/gRPC
+   command reaches straight into `services/`, that's a smell.
+3. It is one of three shapes: a registry/resolver threaded through `Ctx`,
+   an event listener reacting across domains, or a pure algorithm with no
+   persistence.
+
+The test: if you can name one domain whose commands would break without the
+file and no other, it's not a shared service — move it.
+
+Current shared services (and why each stays):
+
+- `capability` — capability registry held on `Ctx`; used by every domain.
+- `event` — event persistence + fanout; called by multiple emitting domains.
+- `llm_resolver` — spans `llm_providers`, `llm_models`, and request params;
+  no single owner.
+- `model_sync` — background listener reconciling provider catalogs with the
+  models table.
+- `principal` — resolves users and agent identities; shared lookup helper.
+- `usage_tracking` — event listener feeding budgets across domains.
+
+Everything else belongs under `domains/<owner>/`. For example, `eval_runner`
+lives at `domains/evals/runner.rs`, `scoped_mcp` at
+`domains/mcp_servers/scoped_mcp.rs`, `virtual_mount_registry` at
+`domains/session_files/virtual_mount_registry.rs`,
+`capability_validation` at `domains/capabilities/validation.rs`,
+`leased_resource` at `domains/session_resources/leased_resource.rs`.
 
 ## Testing
